% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ps_calc.R
\name{ape_calc}
\alias{ape_calc}
\title{Weighted mean different}
\usage{
ape_calc(weights, treatment, outcome, n_bootstrap = 200)
}
\arguments{
\item{weights}{weights}

\item{treatment}{treatment vector}

\item{outcome}{outcome vector}

\item{n_bootstrap}{bootstrap samples}
}
\value{
object with dataframe with LR PS calculation for to balance each
independent varaible
}
\description{
Weighted mean different
}
\examples{
ape_calc(rnorm(100,5,0.5),
sample(c(0,1), 100, replace = TRUE),
sample(c(0,1), 100, replace = TRUE))
}
