% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart_preprocess}
\alias{heart_preprocess}
\title{Heart dataset example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1025 rows and 15 columns.
}
\usage{
heart_preprocess
}
\description{
Heart dataset example
}
\keyword{datasets}
