% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj1}
\alias{obj1}
\title{Object returned from preprocess_daata}
\format{
An object of class \code{list} of length 5.
}
\usage{
obj1
}
\description{
Object returned from preprocess_daata
}
\keyword{datasets}
