% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot.R
\name{plot_forest}
\alias{plot_forest}
\title{Plot Forest
Plot a forest plot to represent}
\usage{
plot_forest(obj)
}
\arguments{
\item{obj}{object from ps_calc()}
}
\value{
object with plot representing the results
}
\description{
Plot Forest
Plot a forest plot to represent
}
\examples{
plot_forest(obj2)
}
