% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data to be analysed}
\usage{
preprocess_data(
  data = heart_preprocess,
  indeps = NULL,
  y = "Heart Disease",
  pred = "predictions"
)
}
\arguments{
\item{data}{Model data}

\item{indeps}{String representing independent variables}

\item{y}{String representing Real outcome}

\item{pred}{String representing Model Predictions}
}
\value{
list with tibble with independent variables, model predictions and
real outcome
}
\description{
Preprocess data to be analysed
}
\examples{
preprocess_data(data = heart_preprocess,
indeps = NULL,
y = "Heart Disease",
pred = "predictions")
}
