% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ps_calc.R
\name{ps_calc}
\alias{ps_calc}
\title{PS calculation}
\usage{
ps_calc(obj, threshold = 0.1, n_bootstrap = 200)
}
\arguments{
\item{obj}{object returned from preprocess_data()}

\item{threshold}{SMD threshold to be balanced}

\item{n_bootstrap}{bootstrap samples}
}
\value{
object with dataframe with LR PS calculation for to balance each
independent variable
}
\description{
PS calculation
}
\examples{
ps_calc(obj1)
}
